rename gatt to bluetooth and fix some typos
