+++ coralibre-sdk/sdk/src/main/java/org/coralibre/android/sdk/internal/bluetooth/BluetoothServiceStatus.java
@@ -8,7 +8,7 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-package org.coralibre.android.sdk.internal.gatt;
+package org.coralibre.android.sdk.internal.bluetooth;
 
 import android.annotation.SuppressLint;
 import android.bluetooth.le.AdvertiseCallback;
