+++ coralibre-sdk/sdk/src/main/java/org/coralibre/android/sdk/internal/database/ppcp/Database.java
@@ -4,7 +4,7 @@ package org.coralibre.android.sdk.internal.database.ppcp;
 import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
 import org.coralibre.android.sdk.internal.database.ppcp.model.CapturedData;
 import org.coralibre.android.sdk.internal.database.ppcp.model.GeneratedTEK;
-import org.coralibre.android.sdk.internal.database.ppcp.model.IntervallOfCapturedData;
+import org.coralibre.android.sdk.internal.database.ppcp.model.IntervalOfCapturedData;
 
 public interface Database {
     void addCapturedPayload(CapturedData collectedPayload);
@@ -12,7 +12,7 @@ public interface Database {
 
     Iterable<GeneratedTEK> getAllGeneratedTEKs();
     GeneratedTEK getGeneratedTEK(ENNumber interval);
-    Iterable<IntervallOfCapturedData> getAllCollectedPayload();
+    Iterable<IntervalOfCapturedData> getAllCollectedPayload();
 
     void truncateLast14Days();
 }
